ref(api): Move to_valid_int_id to a more central location
