+++ src/sentry/api/endpoints/organization_api_key_details.py
@@ -14,10 +14,10 @@ from sentry.api.bases.organization import (
 )
 from sentry.api.exceptions import ResourceDoesNotExist
 from sentry.api.serializers import serialize
+from sentry.api.utils import to_valid_int_id
 from sentry.conf.server import SENTRY_SCOPES
 from sentry.models.apikey import ApiKey
 from sentry.organizations.services.organization.model import RpcOrganization
-from sentry.workflow_engine.endpoints.utils.ids import to_valid_int_id
 
 
 class ApiKeySerializer(serializers.ModelSerializer):
