+++ src/sentry/integrations/api/endpoints/data_forwarding_details.py
@@ -17,6 +17,7 @@ from sentry.api.base import cell_silo_endpoint
 from sentry.api.bases.organization import OrganizationEndpoint, OrganizationPermission
 from sentry.api.exceptions import ResourceDoesNotExist
 from sentry.api.serializers import serialize
+from sentry.api.utils import to_valid_int_id
 from sentry.apidocs.constants import RESPONSE_BAD_REQUEST, RESPONSE_FORBIDDEN, RESPONSE_NO_CONTENT
 from sentry.apidocs.examples.integration_examples import IntegrationExamples
 from sentry.apidocs.parameters import DataForwarderParams, GlobalParams
@@ -35,7 +36,6 @@ from sentry.organizations.services.organization.model import (
     RpcUserOrganizationContext,
 )
 from sentry.web.decorators import set_referrer_policy
-from sentry.workflow_engine.endpoints.utils.ids import to_valid_int_id
 
 
 class OrganizationDataForwardingDetailsPermission(OrganizationPermission):
